fix(go2cs): box an untyped int constant as nint when converted to an interface (Phase 4)
An untyped int constant implicitly converted to an empty interface (`any`) rendered as a bare
C# integer literal — `System.Int32` — but Go boxes an `int` as its default dynamic type `int`
(go2cs `nint`). A later `x.(int)`, emitted `x._<nint>()`, then found a boxed `Int32` and panicked
at runtime ("interface conversion: interface {} is int, not int" — both sides print "int", but one
is Int32, one is nint). fmt masked it (Int32/nint format identically); the bug bit value/type-
assertion round-trips through an interface — e.g. `atomic.Value.Store(42)` then `v.Load().(int)`.

Cast such constants through nint at every empty-interface boxing position — call argument,
var-spec, assignment (including index/selector targets), return, channel send, slice/array element,
and keyed/positional struct field and map value — mirroring the existing string->@string
(castToGoString) family. Driven by argBoxesAsInt32ButNeedsNint (keyed off info.Types, so it catches
`42`, `-5`, `1 + 2`, and named untyped-int consts; excludes defined-type-over-int and values outside
int32 range, which already box correctly). Call arguments reuse the per-argument castArgToType["nint"]
plumbing; the other positions wrap through the new boxUntypedIntAsNint helper.

Three deliberate exclusions:
- variadic ...any (fmt/print/log): a boxed Int32 formats identically and its %T already resolves as
  "int" (golib maps both Int32 and IntPtr to "int"), so the cast is redundant noise on the most
  common call pattern; a non-variadic any parameter (atomic.Value.Store, context.WithValue) is cast.
- type-parameter params constrained by any (func f[T any](v T)): the instantiation binds the argument
  to the concrete type, where a bare int literal converts implicitly (gated via isEmptyInterfaceTarget,
  unlike the u8-span->@string case where a K=string parameter genuinely needs the cast).
- any map KEYS: golib's map uses the default Dictionary comparer (nint(6) != Int32(6)) and index
  lookups are not boxed, so casting the composite key while looking it up as Int32 would break
  map[any]int{6:1}[6] round-trips. Map values ARE cast (retrieved, not compared).

Guarded by the new UntypedIntInterfaceBox behavioral test — each position read back through an
x.(int) assertion or an int/int32 type switch, output-compared vs Go. Re-baselined 12 existing
goldens whose any-boxed int literals now carry the cast (runtime output unchanged). Docs:
ConversionStrategies.md and ConversionStrategies-Reference.md.

Found during Phase-4 operational work while adding the AtomicValue behavioral test (which was
switched to typed variables to sidestep this).

+++ src/Tests/Behavioral/MethodlessFuncTypeAssert/MethodlessFuncTypeAssert.cs
@@ -19,7 +19,7 @@ internal static void Main() {
     any i = new Func<nint, nint>((nint w) => w * 2);
     var c = lookup(i);
     fmt.Println(c(21));
-    any j = 7;
+    any j = (nint)(7);
     fmt.Println(lookup(j) == default!);
 }
